+++ src/lib/function.ts
@@ -1,12 +1,12 @@
 /**
  * Does nothing when called.
  */
-export const noop: (...args: any[]) => any = () => {}; // eslint-disable-line @typescript-eslint/no-empty-function
+export const noop: (...args: Array<any>) => any = () => {}; // eslint-disable-line @typescript-eslint/no-empty-function
 
 /**
  * Async function that returns a resolved promise.
  */
-export const noop_async: (...args: any[]) => Promise<any> = () => resolved;
+export const noop_async: (...args: Array<any>) => Promise<any> = () => resolved;
 
 /**
  * A singleton resolved `Promise`.
@@ -18,11 +18,15 @@ export const resolved = Promise.resolve();
  */
 export const identity = <T>(t: T): T => t;
 
-export type Lazy<T> = () => T;
+/**
+ * Represents a value wrapped in a function.
+ * Useful for lazy values and bridging reactive boundaries in Svelte.
+ */
+export type Thunk<T> = () => T;
 
 /**
  * Returns the result of calling `value` if it's a function,
  * otherwise it's like the `identity` function and passes it through.
  */
-export const lazy = <T>(value: T | Lazy<T>): T =>
-	typeof value === 'function' ? (value as any)() : value; // TODO rename? `call_if_function` something?
+export const unthunk = <T>(value: T | Thunk<T>): T =>
+	typeof value === 'function' ? (value as any)() : value;