+++ src/lib/types.ts
@@ -8,6 +8,10 @@ TODO probably make this `.d.ts` when we make a proper build process
 
 */
 
+export type Class_Constructor<T_Instance, T_Args extends Array<any> = Array<any>> = new (
+	...args: T_Args
+) => T_Instance;
+
 export type Omit_Strict<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
 
 /**