+++ CHANGELOG.md
@@ -1,5 +1,96 @@
 # @ryanatkn/belt
 
+## 0.30.1
+
+### Patch Changes
+
+- upgrade esm-env@1.2.2 ([#40](https://github.com/ryanatkn/belt/pull/40))
+
+## 0.30.0
+
+### Minor Changes
+
+- rename `Thunk` from `Lazy` and `unthunk` from `lazy` ([5e36e11](https://github.com/ryanatkn/belt/commit/5e36e11))
+
+## 0.29.1
+
+### Patch Changes
+
+- add `Class_Constructor` to `types.ts` ([dc8f0b9](https://github.com/ryanatkn/belt/commit/dc8f0b9))
+
+## 0.29.0
+
+### Minor Changes
+
+- change `is_editable` to match all inputs except hidden ones ([4889274](https://github.com/ryanatkn/belt/commit/4889274))
+
+### Patch Changes
+
+- add `is_interactive` ([4889274](https://github.com/ryanatkn/belt/commit/4889274))
+
+## 0.28.0
+
+### Minor Changes
+
+- disable colors by default for better runtime compat - ([1972e95](https://github.com/ryanatkn/belt/commit/1972e95))
+  call `set_colors(styleText)` from `print.ts`,
+  in Node `import {styleText} from 'node:util';`
+
+## 0.27.6
+
+### Patch Changes
+
+- add `strip_ansi` string util ([ffb7587](https://github.com/ryanatkn/belt/commit/ffb7587))
+
+## 0.27.5
+
+### Patch Changes
+
+- export `st` from `print.ts` ([95fe446](https://github.com/ryanatkn/belt/commit/95fe446))
+
+## 0.27.4
+
+### Patch Changes
+
+- tweak logging outputs again ([2cc4100](https://github.com/ryanatkn/belt/commit/2cc4100))
+- tweak logging outputs once more ([2e2fb84](https://github.com/ryanatkn/belt/commit/2e2fb84))
+
+## 0.27.3
+
+### Patch Changes
+
+- tweak logging outputs ([d292845](https://github.com/ryanatkn/belt/commit/d292845))
+
+## 0.27.2
+
+### Patch Changes
+
+- change the debug char ([6876ebc](https://github.com/ryanatkn/belt/commit/6876ebc))
+
+## 0.27.1
+
+### Patch Changes
+
+- improve log symbols and consistency ([28371a1](https://github.com/ryanatkn/belt/commit/28371a1))
+
+## 0.27.0
+
+### Minor Changes
+
+- add `esm-env` optional dep and default logger to debug ([9c8d8af](https://github.com/ryanatkn/belt/commit/9c8d8af))
+
+## 0.26.1
+
+### Patch Changes
+
+- improve `Json` types, adding `Json_Primitive`, `Json_Object`, and `Json_Array` ([9db9a1d](https://github.com/ryanatkn/belt/commit/9db9a1d))
+
+## 0.26.0
+
+### Minor Changes
+
+- remove `obtainable` ([#39](https://github.com/ryanatkn/belt/pull/39))
+
 ## 0.25.3
 
 ### Patch Changes